Added alternative fueling stations to the map
