+++ data/Fuel_Corridors/README.md
@@ -1,0 +1,3 @@
+# Fueling Stations for Alternative Fueling Corridors
+
+The shapefiles for these alternative fueling stations were obtained from https://afdc.energy.gov/corridors
