+++ data/Fuel_Corridors/US_hy/abbr_fields_data_dictionary.csv
@@ -1,0 +1,49 @@
+abbreviated_field_name,full_field_name
+access_cod,access_code
+access_d_1,access_days_time
+access_d_2,access_days_time_fr
+access_det,access_detail_code
+bd_blends,bd_blends_fr
+cards_acce,cards_accepted
+cng_dispen,cng_dispenser_num
+cng_fill_t,cng_fill_type_code
+cng_renewa,cng_renewable_source
+cng_totl_1,cng_total_compression
+cng_totl_2,cng_total_storage
+cng_vehicl,cng_vehicle_class
+date_last,date_last_confirmed
+expected_d,expected_date
+e85_blende,e85_blender_pump
+e85_other,e85_other_ethanol_blends
+ev_connect,ev_connector_types
+ev_dc_fast,ev_dc_fast_num
+ev_level1,ev_level1_evse_num
+ev_level2,ev_level2_evse_num
+ev_netwo_1,ev_network_web
+ev_netwo_2,ev_network_ids
+ev_other_e,ev_other_evse
+ev_prici_1,ev_pricing_fr
+ev_renewab,ev_renewable_source
+facility_t,facility_type
+federal_ag,federal_agency
+fuel_type_,fuel_type_code
+geocode_st,geocode_status
+groups_w_1,groups_with_access_code
+groups_w_2,groups_with_access_code_fr
+hy_is_reta,hy_is_retail
+hy_pressur,hy_pressures
+hy_standar,hy_standards
+hy_status,hy_status_link
+inter_dr_1,intersection_directions
+inter_dr_2,intersection_directions_fr
+lng_renewa,lng_renewable_source
+lng_vehicl,lng_vehicle_class
+lpg_nozzle,lpg_nozzle_types
+lpg_primar,lpg_primary
+ng_fill_ty,ng_fill_type_code
+ng_vehicle,ng_vehicle_class
+owner_type,owner_type_code
+station_na,station_name
+station_ph,station_phone
+status_cod,status_code
+street_add,street_address
