+++ source/CommonTools.py
@@ -1,0 +1,21 @@
+from pathlib import Path
+import os
+
+def get_top_dir():
+    '''
+    Gets the path to the top level of the git repo (one level up from the source directory)
+        
+    Parameters
+    ----------
+    None
+
+    Returns
+    -------
+    top_dir (string): Path to the top level of the git repo
+        
+    NOTE: None
+    '''
+    source_path = Path(__file__).resolve()
+    source_dir = source_path.parent
+    top_dir = os.path.dirname(source_dir)
+    return top_dir
